requirements.txt
alembic==1.5.7
certifi==2020.12.5
chardet==4.0.0
idna==2.10
loguru==0.5.3
Mako==1.1.4
MarkupSafe==1.1.1
psycopg2-binary==2.8.6
pyTelegramBotAPI==3.7.6
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.23
urllib3==1.26.3
psycopg2
alembic
python-telegram-bot-pagination
telebot-calendar==1.1
jinja2
weasyprint
